+++ crates/matrix-sdk-ui/src/room_list_service/filters/space.rs
@@ -31,7 +31,7 @@ pub fn new_filter() -> impl Filter {
 
 #[cfg(test)]
 mod tests {
-    use matrix_sdk::test_utils::logged_in_client_with_server;
+    use matrix_sdk::test_utils::mocks::MatrixMockServer;
     use matrix_sdk_test::async_test;
     use ruma::room_id;
 
@@ -39,7 +39,8 @@ mod tests {
 
     #[async_test]
     async fn test_space() {
-        let (client, server) = logged_in_client_with_server().await;
+        let server = MatrixMockServer::new().await;
+        let client = server.client_builder().build().await;
         let [room] = new_rooms([room_id!("!a:b.c")], &client, &server).await;
 
         assert!(!matches(|_| false, &room));