+++ crates/matrix-sdk-ui/tests/integration/room_list_service.rs
@@ -5,17 +5,10 @@ use eyeball_im::VectorDiff;
 use futures_util::{FutureExt, StreamExt, pin_mut};
 use matrix_sdk::{
     Client, RoomDisplayName,
-    config::RequestConfig,
-    test_utils::{
-        logged_in_client_with_server,
-        mocks::{MatrixMockServer, RoomMessagesResponseTemplate},
-        set_client_session, test_client_builder,
-    },
+    test_utils::mocks::{MatrixMockServer, RoomMessagesResponseTemplate},
 };
 use matrix_sdk_base::sync::UnreadNotificationsCount;
-use matrix_sdk_test::{
-    ALICE, async_test, event_factory::EventFactory, mocks::mock_encryption_state,
-};
+use matrix_sdk_test::{ALICE, async_test, event_factory::EventFactory};
 use matrix_sdk_ui::{
     RoomListService,
     room_list_service::{
@@ -31,19 +24,15 @@ use ruma::{
     owned_mxc_uri, room_id,
     time::{Duration, Instant},
 };
-use serde_json::json;
 use stream_assert::{assert_next_matches, assert_pending};
 use tempfile::TempDir;
 use tokio::{spawn, sync::Barrier, task::yield_now, time::sleep};
-use wiremock::{
-    Mock, MockServer, ResponseTemplate,
-    matchers::{header, method, path},
-};
 
 use crate::timeline::sliding_sync::{assert_timeline_stream, timeline_event};
 
-async fn new_room_list_service() -> Result<(Client, MockServer, RoomListService), Error> {
-    let (client, server) = logged_in_client_with_server().await;
+async fn new_room_list_service() -> Result<(Client, MatrixMockServer, RoomListService), Error> {
+    let server = MatrixMockServer::new().await;
+    let client = server.client_builder().build().await;
     let room_list = RoomListService::new(client.clone()).await?;
 
     Ok((client, server, room_list))
@@ -51,15 +40,10 @@ async fn new_room_list_service() -> Result<(Client, MockServer, RoomListService)
 
 async fn new_persistent_room_list_service(
     store_path: &std::path::Path,
-) -> Result<(MockServer, RoomListService), Error> {
-    let server = MockServer::start().await;
-    let client = test_client_builder(Some(server.uri().to_string()))
-        .request_config(RequestConfig::new().disable_retry())
-        .sqlite_store(store_path, None)
-        .build()
-        .await
-        .unwrap();
-    set_client_session(&client).await;
+) -> Result<(MatrixMockServer, RoomListService), Error> {
+    let server = MatrixMockServer::new().await;
+    let client =
+        server.client_builder().on_builder(|b| b.sqlite_store(store_path, None)).build().await;
 
     let room_list = RoomListService::new(client.clone()).await?;
 
@@ -68,8 +52,9 @@ async fn new_persistent_room_list_service(
 
 async fn new_room_list_service_with_all_rooms_required_state(
     all_rooms_required_state: Vec<(StateEventType, String)>,
-) -> Result<(MockServer, RoomListService), Error> {
-    let (client, server) = logged_in_client_with_server().await;
+) -> Result<(MatrixMockServer, RoomListService), Error> {
+    let server = MatrixMockServer::new().await;
+    let client = server.client_builder().build().await;
     let room_list =
         RoomListService::new_with_all_rooms_required_state(client, true, all_rooms_required_state)
             .await?;
@@ -390,6 +375,7 @@ async fn test_sync_all_states() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                     ],
                     "filters": {},
                     "timeline_limit": 1,
@@ -577,6 +563,7 @@ async fn test_all_rooms_required_state_is_added_to_all_rooms_only() -> Result<()
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.guest_access", ""],
                     ],
                     "filters": {},
@@ -675,6 +662,7 @@ async fn test_all_rooms_required_state_does_not_affect_room_subscriptions() -> R
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.guest_access", ""],
                     ],
                     "filters": {},
@@ -700,6 +688,7 @@ async fn test_all_rooms_required_state_does_not_affect_room_subscriptions() -> R
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.pinned_events", ""],
                     ],
                     "timeline_limit": 20,
@@ -757,6 +746,7 @@ async fn test_all_rooms_required_state_overrides_default_event_type_entries() ->
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.member", "*"],
                     ],
                     "filters": {},
@@ -2613,6 +2603,7 @@ async fn test_room_subscription() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.pinned_events", ""],
                     ],
                     "timeline_limit": 20,
@@ -2655,6 +2646,7 @@ async fn test_room_subscription() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                     ],
                     "filters": {},
                     "timeline_limit": 1,
@@ -2679,6 +2671,7 @@ async fn test_room_subscription() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.pinned_events", ""],
                     ],
                     "timeline_limit": 20,
@@ -2726,6 +2719,7 @@ async fn test_room_subscription() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                     ],
                     "filters": {},
                     "timeline_limit": 1,
@@ -2750,6 +2744,7 @@ async fn test_room_subscription() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.pinned_events", ""],
                     ],
                     "timeline_limit": 20,
@@ -2772,6 +2767,7 @@ async fn test_room_subscription() -> Result<(), Error> {
                         ["io.element.functional_members", ""],
                         ["m.space.parent", "*"],
                         ["m.space.child", "*"],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                         ["m.room.pinned_events", ""],
                     ],
                     "timeline_limit": 20,
@@ -2895,7 +2891,7 @@ async fn test_room_timeline() -> Result<(), Error> {
         },
     };
 
-    mock_encryption_state(&server, false).await;
+    server.mock_room_state_encryption().plain().mount().await;
 
     let room = room_list.room(room_id)?;
     let timeline = room.timeline_builder().build().await.unwrap();
@@ -2946,16 +2942,9 @@ async fn test_room_timeline() -> Result<(), Error> {
 #[async_test]
 async fn test_room_empty_timeline() {
     let (client, server, room_list) = new_room_list_service().await.unwrap();
-    mock_encryption_state(&server, false).await;
+    server.mock_room_state_encryption().plain().mount().await;
 
-    Mock::given(method("POST"))
-        .and(path("_matrix/client/r0/createRoom"))
-        .and(header("authorization", "Bearer 1234"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(json!({ "room_id": "!example:localhost"})),
-        )
-        .mount(&server)
-        .await;
+    server.mock_create_room().ok().mount().await;
 
     let room = client.create_room(CreateRoomRequest::default()).await.unwrap();
     let room_id = room.room_id().to_owned();
@@ -2973,7 +2962,7 @@ async fn test_room_empty_timeline() {
 #[async_test]
 async fn test_room_latest_event() -> Result<(), Error> {
     let (client, server, room_list) = new_room_list_service().await?;
-    mock_encryption_state(&server, false).await;
+    server.mock_room_state_encryption().plain().mount().await;
 
     let sync = room_list.sync();
     pin_mut!(sync);
@@ -3363,6 +3352,7 @@ async fn test_thread_subscriptions_extension_enabled_only_if_server_advertises_i
                             [ "io.element.functional_members", "", ],
                             [ "m.space.parent", "*", ],
                             [ "m.space.child", "*", ],
+                            ["org.matrix.msc3672.beacon_info", "*"],
                         ],
                         "timeline_limit": 1,
                     },
@@ -3460,6 +3450,7 @@ async fn test_thread_subscriptions_extension_enabled_only_if_server_advertises_i
                         [ "io.element.functional_members", "", ],
                         [ "m.space.parent", "*", ],
                         [ "m.space.child", "*", ],
+                        ["org.matrix.msc3672.beacon_info", "*"],
                     ],
                     "timeline_limit": 1,
                 },